requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
urllib3==2.6.3
chromadb==0.5.5
tqdm==4.66.5
openai>=1.0.0
python-dotenv>=1.0.0